Refactor/tico back end

+++ src/main/java/com/femcoders/tico/dto/response/UserResponseDTO.java
@@ -9,7 +9,7 @@ public record UserResponseDTO(
         String email,
         Set<String> roles,
         Boolean isActive,
-        int openTickets,
+        long openTickets,
         LocalDateTime createdAt
         ) {
